requirements_minimal.txt
# Минимальные зависимости для базовой работы
Flask==2.3.3
Werkzeug==2.3.7
psutil==5.9.6
py-cpuinfo==9.0.0
